+++ src/grandorgue/GOrgueSoundPortaudioPort.h
@@ -33,6 +33,7 @@ private:
 	static int Callback (const void *input, void *output, unsigned long frameCount, const PaStreamCallbackTimeInfo *timeInfo, PaStreamCallbackFlags statusFlags, void *userData);
 
 	static wxString getName(unsigned index);
+	static wxString getLastError(PaError error);
 
 public:
 	static const wxString PORT_NAME;